requirements.txt
fastapi
uvicorn
pydantic
pytube
openai-whisper